+++ src/modules/crm/infrastructure/container.ts
@@ -1,6 +1,7 @@
 import { organizationQueries } from "@/modules/organizations";
 import { makeCrmCommands } from "../application/commands";
 import { makeCrmQueries } from "../application/queries";
+import { makeDetectCrmInsights } from "../application/detect-insights";
 import { makeCustomerDirectory } from "../application/customer-directory";
 import { PrismaCustomerRepository } from "./customer.repository";
 import { PrismaFollowerRepository } from "./follower.repository";
@@ -11,6 +12,7 @@ const followers = new PrismaFollowerRepository();
 /** Composition root for the crm module. */
 export const crmCommands = makeCrmCommands({ customers, followers });
 export const crmQueries = makeCrmQueries({ customers, followers });
+export const detectCrmInsights = makeDetectCrmInsights({ crm: crmQueries });
 
 export const customerDirectory = makeCustomerDirectory({
   organizations: organizationQueries,